k8s_builder/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
python-keycloak==3.0.0
pandas==2.0.2
PyYAML==6.0
openpyxl==3.1.2
pytest==7.3.1
deepdiff==6.3.0